Fixed failing tests because of circular dependency
